// apps/api/src/dtos/auth.dto.d.ts
export declare class LoginDto {
    email: string;
    password: string;
}
export declare class RegisterDto {
    username: string;
    email: string;
    password: string;
    firstName?: string;
    lastName?: string;
}
export declare class TokenDto {
    accessToken: string;
    refreshToken: string;
}